+++ crates/index/test_data/.gitattributes
