+++ crates/index/test_data/100-points.txt
