+++ crates/index/test_data/1000-points.txt
