+++ crates/index/test_data/10000-points.txt
