--- crates/test_utils/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "de_test_utils"
-description = "Test utilities for Digital Extinction"
-
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-repository.workspace = true
-keywords.workspace = true
-homepage.workspace = true
-license.workspace = true
-categories.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-glam.workspace = true
